refactor LdapAuthTest
